chore: upgrade derive_more to 0.99.14
We are using `Display` macro from `derive_more` crate to easily
implement `std::fmt::Display` trait for each lint's message and hint.
This macro was a bit tricky; even if passed arguments are invalid format,
it _does_ compile, which results in getting unexpected diagnostic outputs.
This is why the issue #666 happened.

However, in `derive_more` v0.99.14, the macro has been refined so it now
rejects arguments in invalid format at compile-time. This is so useful
for us not to make a mistake.
Ref: https://github.com/JelteF/derive_more/pull/160