fix: Allow `deno-lint-ignore-file` comments after a shebang
File-level ignore directives were being looked for in the leading
comments before the first token in the program, but when the file has a
shebang, the shebang is the first token, and so no comments are
returned. This change makes it so, if the program has a shebang, the
directives are looked for immediately before the first statement or
declaration, if any, or at the end of the program otherwise.

Closes #999.